Create a standard server-side console logger - Use this in both the main server and the status server.
