+++ status/index.ts
@@ -7,29 +7,10 @@ import {
   serveFile,
 } from "../server/util/serving.ts";
 import { addLogger } from "../lib/log.ts";
-import {
-  isStringWithOptions,
-  makeConsoleLogger,
-  StringWithOptions,
-} from "../lib/console_logger.ts";
+import { makeConsoleLogger } from "../lib/console_logger.ts";
+import { consoleLogger } from "../server/util/console_logger.ts";
 
-addLogger(
-  makeConsoleLogger(
-    (...strs: (string | StringWithOptions)[]) => {
-      // TODO(applmak): Implement actual color here.
-      const coloredStrings: string[] = [];
-      for (const str of strs) {
-        if (isStringWithOptions(str)) {
-          coloredStrings.push(str.str);
-        } else {
-          coloredStrings.push(str);
-        }
-      }
-      console.log(coloredStrings.join(""));
-    },
-    () => performance.now(),
-  ),
-);
+addLogger(makeConsoleLogger(consoleLogger, () => performance.now()));
 
 const FLAG_DEFS = [
   { name: "port", type: Number, defaultValue: 3000 },
